Vid_to_Query-RAG/requirements.txt
numpy
pandas
scikit-learn
joblib
requests
openai-whisper
ffmpeg
bge-m3
llama3.2:latest